+++ src/commands.js
@@ -1,3 +1,12 @@
-module.exports = yargs => {
-    yargs.commandDir('commands')
-}
+const { packageConfigs, styleConfig } = require('./utils/config.js')
+
+module.exports = yargs =>
+    yargs
+        .config({
+            config: {
+                ...require(packageConfigs.d2Style),
+                ...styleConfig,
+            },
+            files: [],
+        })
+        .commandDir('commands')