+++ src/tools/commitlint.js
@@ -1,14 +1,17 @@
-const { COMMITLINT_CONFIG } = require('../utils/paths.js')
+const { packageConfigs } = require('../utils/config.js')
 const { bin } = require('../utils/run.js')
 
-exports.commitlint = ({ config = COMMITLINT_CONFIG, file }) => {
+exports.commitlint = ({
+    config = packageConfigs.commitlint,
+    file,
+    callback,
+}) => {
     const cmd = 'commitlint'
     const args = [
         'commitlint',
-        `--config=${config}`,
-        '--edit',
-        ...(file ? [file] : []),
+        ...(config ? ['--config', config] : []),
+        ...(file ? ['--edit', file] : []),
     ]
 
-    bin(cmd, { args })
+    bin(cmd, { args }, callback)
 }