+++ src/tools/eslint.js
@@ -2,15 +2,13 @@ const { resolveIgnoreFile } = require('../utils/files.js')
 const { PACKAGE_ROOT } = require('../utils/paths.js')
 const { bin } = require('../utils/run.js')
 
-exports.eslint = ({ files = [], apply = false, config }) => {
+exports.eslint = ({ files = [], apply = false, config, callback }) => {
     const ignoreFile = resolveIgnoreFile(['.eslintignore'])
     const cmd = 'eslint'
     const args = [
-        '--no-color',
         '--report-unused-disable-directives',
         '--ignore',
-        '--quiet',
-        '--format=unix',
+        '--format=stylish',
         `--resolve-plugins-relative-to=${PACKAGE_ROOT}`,
         ...(ignoreFile ? ['--ignore-path', ignoreFile] : []),
         ...(config ? ['--config', config] : []),
@@ -18,5 +16,5 @@ exports.eslint = ({ files = [], apply = false, config }) => {
         ...files,
     ]
 
-    bin(cmd, { args })
+    bin(cmd, { args }, callback)
 }