+++ src/utils/run.js
@@ -1,17 +1,20 @@
 const path = require('path')
+const log = require('@dhis2/cli-helpers-engine').reporter
 const spawn = require('cross-spawn')
 const findup = require('find-up')
 const { PACKAGE_ROOT } = require('./paths.js')
 
-exports.spawn = (cmd, args, opts) =>
-    spawn.sync(cmd, args, {
+exports.spawn = (cmd, args, opts) => {
+    log.debug(cmd, args, opts)
+    return spawn.sync(cmd, args, {
         stdio: 'inherit',
         ...opts,
     })
+}
 
 exports.run = (cmd, { args, opts }, callback) => {
     return handleRun(
-        spawn.sync(cmd, args, {
+        this.spawn(cmd, args, {
             stdio: 'inherit',
             ...opts,
         }),
@@ -29,12 +32,31 @@ exports.bin = (cmd, { args, opts }, callback) => {
     const binCmd = path.join(nodemodules, '.bin', cmd)
 
     return handleRun(
-        spawn.sync(binCmd, args, {
+        this.spawn(binCmd, args, {
             stdio: 'inherit',
             ...opts,
         }),
         callback
     )
+}
+
+exports.callback = () => {
+    let status = 0
+    return result => {
+        if (!result) return status
+
+        if (result.status > status) {
+            status = result.status
+        }
+    }
+}
+
+exports.exit = (code, msg) => {
+    if (msg && code > 0) {
+        console.log('')
+        log.error(msg)
+    }
+    process.exit(code)
 }
 
 function handleRun(result, callback) {
@@ -43,7 +65,7 @@ function handleRun(result, callback) {
     }
 
     if (callback) {
-        callback(result)
+        return callback(result)
     }
 
     if (result.status !== 0) {